Add files via upload
a quiz game that asks users multiple-choice on a specific topic. The game should keep track of scores, offer a variety of questions to make it challenging and engaging and user provide feedback to improve the quiz game  .
question and answer -  This quiz game show 5 question and many options are avilable for the answer and the topic of quiz is python . 
Display each question and answer choice.
Prompt the user to select an answer.
score - Calculate the Final Score and Display Final Results. 
Display if the answer was correct or incorrect.
Print the user's score in percentage .