Add files via upload
A password generator is a useful tool that generates strong and random passwords for users. This project aims to create a password generator application using Python, allowing users to specify the length and complexity of the password. and this it prompt the user for length of password  and then generate the password using different combination of characters and symbols and print the password .